requirements.txt
requests~=2.32.3
allure-python-commons~=2.13.5
pydantic~=2.9.2
mimesis~=18.0.0